+++ config/runtime.exs
@@ -27,7 +27,7 @@ if config_env() == :prod do
     monitored_app_name: System.fetch_env!("DEPLOYEX_MONITORED_APP_NAME"),
     monitored_app_log_path: "/var/log",
     phx_start_port: String.to_integer(System.get_env("DEPLOYEX_MONITORED_APP_PORT") || "4000"),
-    replicas: String.to_integer(System.get_env("DEPLOYEX_MONITORED_REPLICAS") || "2")
+    replicas: String.to_integer(System.get_env("DEPLOYEX_MONITORED_REPLICAS") || "3")
 
   config :ex_aws,
     region: System.fetch_env!("AWS_REGION")